Move json to private and delete id.ts
