+++ packages/cli-kit/src/index.ts
@@ -6,7 +6,6 @@ export * as error from './error.js'
 export * as file from './file.js'
 export * as git from './git.js'
 export * as http from './http.js'
-export * as id from './id.js'
 export * as npm from './npm.js'
 export * as os from './os.js'
 export * as output from './output.js'
