+++ packages/cli-kit/src/metadata.ts
@@ -1,5 +1,5 @@
 import {isUnitTest} from './environment/local.js'
-import {AnyJson} from './json.js'
+import {AnyJson} from './private/common/json.js'
 import {MonorailEventPublic} from './monorail.js'
 import {sendErrorToBugsnag} from './public/node/error-handler.js'
 import {PickByPrefix} from './typing/pick-by-prefix.js'
