+++ packages/cli-kit/src/monorail.ts
@@ -1,6 +1,6 @@
 import {fetch} from './http.js'
 import {debug, content, token} from './output.js'
-import {JsonMap} from './json.js'
+import {JsonMap} from './private/common/json.js'
 import {DeepRequired} from './typing/deep-required.js'
 
 const url = 'https://monorail-edge.shopifysvc.com/v1/produce'
