+++ packages/cli-kit/src/plugins.ts
@@ -1,4 +1,4 @@
-import {JsonMap} from './json.js'
+import {JsonMap} from './private/common/json.js'
 import {PickByPrefix} from './typing/pick-by-prefix.js'
 import {MonorailEventPublic, MonorailEventSensitive} from './monorail.js'
 import {HookReturnPerTunnelPlugin} from './public/node/plugins/tunnel.js'
