+++ packages/cli-kit/src/public/node/presets.ts
@@ -1,5 +1,5 @@
 import {exists as fileExists, read as fileRead} from '../../file.js'
-import {JsonMap} from '../../json.js'
+import {JsonMap} from '../../private/common/json.js'
 import {findUp, join as pathJoin} from '../../path.js'
 import {decode as tomlDecode} from '../../toml.js'
